requirements.txt
Bmatplotlib>2.0.0
numpy>=1.0.0
protobuf>=3.0.0
tensorboardX>=1.3
torch>=1.0.1
torchvision
